requirements.txt
fpdf2==2.8.2
Pillow==10.0.1
Pillow==11.1.0
xlrd==2.0.1
